requirements.txt
accelerate==1.9.0
flower==2.0.1
open_clip_torch==3.1.0
pymongo==4.14.0
qdrant-client==1.15.1
transformers==4.55.0
celery==5.5.3
redis==4.6.0
python-dotenv==1.1.1
fastapi == 0.116.1
fastapi-cli == 0.0.8
uvicorn == 0.35.0
requests==2.32.4
